refactor(test): split preview-visual tests by comic document cache and preload behavior
